+++ cookbook/recipes/potential_mag_prism.py
@@ -8,16 +8,17 @@ log = ft.log.get()
 log.info(ft.log.header())
 log.info(__doc__)
 
+bounds = [-5000, 5000, -5000, 5000, 0, 5000]
 prisms = [
     ft.msh.ddd.Prism(-4000,-3000,-4000,-3000,0,2000,
-        {'magnetization':20}),
+        {'magnetization':2}),
     ft.msh.ddd.Prism(-1000,1000,-1000,1000,0,2000,
-        {'magnetization':10}),
+        {'magnetization':1}),
     ft.msh.ddd.Prism(2000,4000,3000,4000,0,2000,
-        {'magnetization':30, 'inclination':-10, 'declination':45})]
+        {'magnetization':3, 'inclination':-10, 'declination':45})]
 # Create a regular grid at 100m height
 shape = (200, 200)
-area = (-5000, 5000, -5000, 5000)
+area = bounds[:4]
 xp, yp, zp = ft.grd.regular(area, shape, z=-500)
 # Calculate the anomaly for a given regional field
 tf = ft.pot.prism.tf(xp, yp, zp, prisms, 30, -15)
@@ -30,3 +31,10 @@ ft.vis.colorbar()
 ft.vis.xlabel('East y (km)')
 ft.vis.ylabel('North x (km)')
 ft.vis.show()
+# Show the prisms
+ft.vis.figure3d()
+ft.vis.prisms(prisms, 'magnetization')
+ft.vis.axes3d(ft.vis.outline3d(bounds), ranges=[i*0.001 for i in bounds])
+ft.vis.wall_north(bounds)
+ft.vis.wall_bottom(bounds)
+ft.vis.show3d()